Refactoring of geometry sorting functions (PMT and CRT).
